code/web-application/docker/requirements-docker.txt
pytest==7.4.0
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
timeout-decorator==0.4.1